backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg2-binary
redis
pydantic
pydantic-settings
requests
beautifulsoup4
pandas
python-dotenv
google-generativeai
streamlit

